refactor!: split the project code into several files(directories)
