+++ .gitignore
@@ -1,5 +1,7 @@
 */.venv
-*/__pycache__
+**/__pycache__/
 /.idea
+todos.txt
+assets
 snippets.txt
 *.bak
