+++ app/graphics/tileset.py
@@ -1,0 +1,17 @@
+import pygame
+
+
+class Tileset:
+  def __init__(self, path, tilesize):
+    self.image = pygame.image.load(path)
+    if self.image.get_alpha():
+      self.image = self.image.convert_alpha()
+    else:
+      self.image = self.image.convert()
+      self.image.set_colorkey(self.image.get_at((0, 0)))
+    self.tilesize = tilesize
+
+  def crop(self, x, y):
+    tsize = [self.tilesize] * 2
+    result = self.image.subsurface(pygame.Rect((x, y), tsize)).copy()
+    return result
